+++ src/components/common/modal.module.scss
@@ -1,0 +1,22 @@
+$modal-background-color: white;
+
+.modal-overlay {
+  position: absolute;
+  top: 0;
+  left: 1;
+  width: 375px;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: rgb(0 0 0 / 30%);
+
+  .modal {
+    display: flex;
+    flex-direction: column;
+    padding: 22px 15px;
+    background: $modal-background-color;
+    width: 300px;
+    height: 350px;
+  }
+}
